refactor(pilot): split _internal file into directory
We can code split things better this way. Plus, it was getting kinda messy with all sorts of different things just being thrown in there.
