+++ packages/pilot/src/_generated-pages.ts
@@ -1,7 +1,7 @@
 /**
  * © 2022 WavePlay <dev@waveplay.com>
  */
-import type { PageModule, PageRoute } from './_internal';
+import type { PageModule, PageRoute } from './_internal/types';
 
 // This is just a stub for what gets automatically generated by `pilot build`
 export const pageRoutes: PageRoute[] = [];
